Common test bit.
